Heavy  Effort on Dialing In the Slick Function
Got SCSS Compiling So I could Implement Text Overlay. After i Got Overlay working I started Modifying and improving the function to generate Slick Blocks. Now I Can In the Custom Block Template Call and Make the slick slider a grid, lazy load. I will be removing the VW/VH from it. As I do not want to deal with that on such a low level. But it made it easier to fix the JS ect. Next Major feature will be Adding Some Low Level JS Cropping/ Margin for the img's. This is to allow the function to be used in as many scenarios as possible.