Found a Damn useful CURL Script for grabbing reviews
I need to take the time to re-write this curl call from ground up. I need Images for users as well. But this guys script is a sold start. Just need to Work backwards and grab user img's with reviews. I gotta check the Json array returned. for now the script is inside of a new block php template reviews.